relocate config and logging
