+++ template/src/{{ project_slug }}/__init__.py
@@ -1,15 +1,2 @@
-from ._config import project_settings, settings
-from ._logging import logger
-from .core import paths
-
-
-__all__ = [
-    "logger",
-    "paths",
-    "settings",
-    "project_settings",
-]
-
-
 # Your code goes here (add your code between this line and the EOF comment)
 # End of file
