+++ template/src/{{ project_slug }}/context.py
@@ -1,6 +1,6 @@
 from pydantic import UUID4, BaseModel
 
-from ._config import project_settings, settings
+from .core import project_settings, settings
 
 
 class Context(BaseModel):
