+++ template/src/{{ project_slug }}/core/__init__.py
@@ -1,0 +1,12 @@
+import paths
+
+from .config import project_settings, settings
+from .logging import logger
+
+
+__all__ = [
+    "logger",
+    "paths",
+    "settings",
+    "project_settings",
+]
