+++ template/src/{{ project_slug }}/core/logging.py
