+++ template/src/{{ project_slug }}/frontend/email.py
@@ -1,7 +1,7 @@
 from pydantic import EmailStr
 from starlette_babel import gettext_lazy as _
 
-from .._config import project_settings
+from ..core import project_settings
 from ..services.email import SESEmailService
 from .templates import render_static_template
