+++ template/src/{{ project_slug }}/services/email.py
@@ -2,7 +2,7 @@ from typing import Literal
 
 import boto3
 
-from .._config import project_settings, settings
+from ..core import project_settings, settings
 
 
 SES_SERVICE_NAME: Literal["ses"] = "ses"
